+++ python/aitemplate/compiler/transform/transform_memory_ops.py
@@ -16,14 +16,15 @@
 Perform memory operator related transformations.
 """
 import copy
-from typing import List, Optional
+from typing import List
 
-from aitemplate.compiler.base import IntImm, IntVar, Operator, Tensor
+from aitemplate.compiler.base import Operator, Tensor
 
-from aitemplate.compiler.ops.tensor.dynamic_slice import dynamic_slice, MAX_INT32
+from aitemplate.compiler.ops.tensor.dynamic_slice import dynamic_slice
 from aitemplate.compiler.tensor_accessor import TensorAccessor
 from aitemplate.compiler.transform import transform_strided_ops_utils, transform_utils
 from aitemplate.compiler.transform.toposort import toposort
+from aitemplate.compiler.transform.transform_merge_slice_ops import merge_slice_ops
 
 from aitemplate.utils import graph_utils, shape_utils
 
@@ -86,7 +87,7 @@ def _update_cat_dst_ops(
         else:
             # Make a new slice op. Note that it's fine we make a new slice op from
             # another slice op, because consecutive slice ops will be merged
-            # by the _try_merge_slice_slice pass
+            # by the merge_slice_ops pass
             slice_start_indices = [0] * rank
             slice_end_indices = [None] * rank
             slice_start_indices[cat_dim] = cat_dim_offset
@@ -420,119 +421,6 @@ def _merge_split_and_cat(sorted_graph: List[Tensor]) -> List[Tensor]:  # noqa: C
     return transform_utils.sanitize_sorted_graph(sorted_graph)
 
 
-def _try_merge_slice_slice(
-    first_slice: Operator, second_slice: Operator, slice_dim: int
-) -> bool:
-    """
-    This function tries to merge two consecutive slice ops with the following
-    steps:
-        * update the start_indices and end_indices fields of the second_slice
-        * remove the first slice
-    """
-    first_slice_output = first_slice._attrs["outputs"][0]
-    first_slice_input_shape = first_slice._attrs["inputs"][0].shape()
-    second_slice_output = second_slice._attrs["outputs"][0]
-    second_slice_output_shape = second_slice_output.shape()
-    # note that all the dims of input_shape[slice_dim:] and output_shape[slice_dim:]
-    # are static at this point
-    for idx in range(slice_dim, first_slice_output._rank()):
-        first_slice_dim_offset = first_slice._attrs["start_indices"][idx]
-        # update the start and end indices of the second slice op
-        new_start = second_slice._attrs["start_indices"][idx] + first_slice_dim_offset
-        first_slice_input_dim = first_slice_input_shape[idx].value()
-        # new start index exceeds the corresponding dim value of the first slice input shape
-        if new_start >= first_slice_input_dim:
-            return False
-        new_end = new_start + second_slice_output_shape[idx].value()
-        # new end index exceeds the corresponding dim value of the first slice input shape
-        if new_end > first_slice_input_dim:
-            return False
-        first_slice_end = first_slice._attrs["end_indices"][idx]
-        second_slice_end = second_slice._attrs["end_indices"][idx]
-        if first_slice_end == MAX_INT32 == second_slice_end:
-            new_end = MAX_INT32
-        second_slice._attrs["start_indices"][idx] = new_start
-        second_slice._attrs["end_indices"][idx] = new_end
-    # remove the old strided op from the first cat's dst_ops
-    transform_utils.remove_single_tensor_op_from_sorted_graph(first_slice)
-    return True
-
-
-def _check_slice_op(slice_op: Operator, slice_dim: int) -> bool:
-    """
-    Return True if the slice_op's indices are valid for being merged
-    """
-    slice_shape = slice_op._attrs["outputs"][0].shape()
-    if not shape_utils.all_static_dimensions(slice_shape, slice_dim):
-        return False
-    # we expect normalized start_indices and end_indices
-    start_index = slice_op._attrs["start_indices"][slice_dim]
-    if start_index is None or start_index < 0:
-        return False
-    end_index = slice_op._attrs["end_indices"][slice_dim]
-    if end_index is None or end_index < 0 or end_index <= start_index:
-        return False
-    return True
-
-
-def _get_rightmost_non_dynamic_dim(shape: List[IntVar]) -> Optional[int]:
-    """
-    Return the index of the rightmost non-dynamic dim. For example, given
-    a shape [3, dyn_dim, 4, 1], it would return 2, which is the index of the
-    third dim.
-    Return None if shape[-1] is dynamic.
-    """
-    idx = 0
-    for dim in reversed(shape):
-        if not isinstance(dim, IntImm):
-            break
-        idx += 1
-    if idx == 0:
-        return None
-    return len(shape) - idx
-
-
-def _merge_slice_and_slice(sorted_graph: List[Tensor]) -> List[Tensor]:
-    # a list of tuple(first_slice, second_slice, slice_dim)
-    to_be_merged = []
-    for tensor in sorted_graph:
-        src_ops = tensor._attrs["src_ops"]
-        if len(src_ops) != 1:
-            continue
-        src_op = list(src_ops)[0]
-        if src_op._attrs["op"] != "dynamic_slice":
-            continue
-        first_slice = src_op
-        first_slice_output = first_slice._attrs["outputs"][0]
-        if first_slice_output._attrs["is_output"]:
-            continue
-        slice_dim = _get_rightmost_non_dynamic_dim(first_slice_output.shape())
-        if slice_dim is None:
-            continue
-        if not _check_slice_op(first_slice, slice_dim):
-            continue
-        next_ops = first_slice_output._attrs["dst_ops"]
-        if len(next_ops) != 1:
-            continue
-        next_op = next_ops[0]
-        if next_op._attrs["op"] != "dynamic_slice":
-            continue
-        second_slice = next_op
-        second_slice_output = second_slice._attrs["outputs"][0]
-        if first_slice_output._rank() != second_slice_output._rank():
-            continue
-        second_slice_dim = _get_rightmost_non_dynamic_dim(second_slice_output.shape())
-        if slice_dim != second_slice_dim:
-            continue
-        if not _check_slice_op(second_slice, slice_dim):
-            continue
-        to_be_merged.append([first_slice, second_slice, slice_dim])
-
-    for first_slice, second_slice, slice_dim in to_be_merged:
-        _try_merge_slice_slice(first_slice, second_slice, slice_dim)
-    return transform_utils.sanitize_sorted_graph(sorted_graph)
-
-
 def _eliminate_split_full_idx(sorted_graph: List[Tensor]) -> List[Tensor]:
     for tensor in sorted_graph:
         src_ops = tensor._attrs["src_ops"]
@@ -572,7 +460,7 @@ def transform_memory_ops(
     funcs = [
         _eliminate_split_full_idx,
         _merge_split_and_cat,
-        _merge_slice_and_slice,
+        merge_slice_ops,
         _eliminate_cat,
     ]
     num_ops = None
